Don't use the user-provided ID as a vertex identifier.
The user-provided IDs are meant to be unique only amongst the siblings
in the graph (which is something the user can manage), not graph-wide
(ensuring cross-tenant unique IDs would be close to impossible for
users).

Therefore we're using an autogenerated vertex ID and have an additional
"uid" property that is checked for uniqueness in the proper context.