[HWKINVENT-25] - Introduce update types to the inventory API.
Similarly to the blueprint types which are used for entity creation,
the update types are used when updating the entity.

The explicit update types make it clear what is possible to update on an
entity in a type-safe manner. Also these update types can be used in the
REST api and no further adaptation for REST is necessary.

Thanks to the explicit update types, we can also have the entities truly
immutable. The update types themselves are made immutable, too, and
consistently use the builder pattern to compose the update prior to their
instantiation.

The RelationshipBrowser and RelationshipService in the tinkerpop impl were
refactored to use the more strict generic types correctly.

Some of the tests in the BasicTest of the tinkerpop impl were removed
because they tested conditions that are no longer possible with the
explicit update types (testing that the relationship update updates only
the correct entities based on the traversal position).

The REST API was updated to contain the update endpoints for all types of
entities and to use the new update types instead of the bespoke DAO classes
inside the REST API impl.