[HWKINVENT-10] Resource hierarchy possible. This is rather big (again) but I think provides a good foundation for future changes.
* "owns" relationship renamed to "incorporates" - the name was confusing
for what the relationship really meant

* Semantical rules for relationship creation/deletion are checked for
throughout the codebase - it is no longer possible to "manually" delete
the contains relationship and wreck havoc in the inventory structure.

* All elements now store a canonical path to them. This makes it conversion
of entities from internal representations much cheaper and also makes it
possible to figure out the canonical path to a nested resource efficiently.
Without it a full traversal from the resource up to tenant would be needed
to obtain that information.

* "isParentOf" relationship added for obvious reasons given the purpose of
this patch ;)