[HWKINVENT-31] Metrics and resources can no live under feeds, too.
This adds the possibility for the resources and metrics to be contained
within feeds, too.

This adds the ability to have "feed-local" resource and metric ids
without any special ID requirements. I.e. two feeds can declare
resources with the same ID easily as long as the resources are contained
within the respective feeds in inventory.

The ability to have feedless resources and metrics has been preserved to
enable ad-hoc user-defined resources.

This is the first commit with a draft implementation of the
functionality.

Missing pieces:
[ ] unit tests
[ ] REST API updates