[HWKINVENT-36] "AutoTenant" support.
This is quite a big refactoring introducing a concept of "mixins" to
the basic inventory - currently "Observable" and "AutoTenant".

Observable is the former ObservableInventory which has been made package
private and made available only through the aforementioned mixin
interface.

The second is AutoTenant, which takes care of automatic creation of
previously unknown tenants. This is in support for the REST API which
will transition the tenant handling to more of the way Hawkular
Metrics are doing it.