Removing READ operations. Under current security model READ is implied.
If the user has access to the tenant, which is implicit by having the
persona available, there's no need check for any READ permissions.

Having such permissions would sooner or later lead to the necessity for
filtering which is something we don't want yet - the security model
will need more thought before we can introduce those.