[HWKINVENT-46] Make event handling error-resilient.
This required a little bit of hackery on rxjava but I think it is the
correct way of doing it in the face of possibly buggy observers.

One could argue that it is the observer's responsibility to obey a contract
(not throw exception in onNext() ) but on the other hand, it is the library
author's responsibility to handle buggy "inputs".