[HWKINVENT-10] Pervasive use of paths instead of ids in the API
So far, we've used a simple ID in the association interface and elsewhere
because there was an implicit assumption that the associated entity is
uniquely identifiable just by its id (and not a full path). We could afford
this assumption because generally speaking it made sense - rarely you would
add a metric type coming from another tenant, or add a metric to resource
coming from another environment or feed.

With the addition of child resources, this assumption no longer holds
because it is assumed that people will want to create alternative tree
views of their resources (to create alternate "groupings" of them based on
external criteria). To be able to associate resources from different
environment or indeed from two different feeds in the same environment, one
needs to be able to address them using a path, not just an id.

Many of the *Read interfaces have been split into 2 - ReadContained and
Read where the former is used to traverse the contains hierarchies which
are by definition still addressable by a simple string. This makes writing
the traversal call chains concise and meaningful in that context.

The latter, Read interfaces, address the traversal targets by a path. This
is to emphasize the fact that id is not enough to address the target entity
because there might be multiple entities with the same id (each in
different contains-hierarchy) on "the other hand" of the traversal step.