[HWKINVENT-106] Meaningful error when parent not found. Instead of IOOBE we now throw an EntityNotFoundException pointing to the parent.
Also, made the TraversalContext store reference to the "source" or previous
traversal context, so that we can access information about the previous
traversal steps. This is currently used only for error reporting in case
of missing parent but I imagine it could be used in other places, too.

Note that the contexts are short lived objects that only live during the
lifetime of a traversal, so it is IMHO not crucial to keep their numbers
at minimum.