[HWKINVENT-128] Operation types
This introduces the operation types as entities contained in resource
types.

Unlike metric types that are merely associated with the resource types but
otherwise are independent, the operation types are truly contained within
resource types and cannot live without it.

The thinking behind that is that unlike metrics, operations really only
make sense on resources. It is not particularly useful to think about
resourceless operations executed directly by the feed. Or rather the cost
of having such standalone operations and having to address them uniquely
within the feed or globally in a tenant outweighs the benefit of having
them.

As such the operation types are unique to the resource types.

WARNING: this commit does NOT address operation types in the REST API.