Redo identity hash computation so that the single algorithm is used.
Before this, the algo was duplicated for MetadataPack.Members and
Entity which was not ideal given the fact that the algorithm is
going to get much more complicated for Merkle tree computation of
many more entity types.

Now a "HashableView" is constructed from either the MT members or
the entity and its dependent entities and this view is then used
to compute the hash.