Totally reworked pre-commit actions so that it is possible to only compose the notifications at the actual commit of the transaction even in the case of a transaction frame.
This is needed to minimize the number of times we recompute the tree
hash in case of bulk insert, etc.

The transaction is no longer subclassable in backends but instead
they can attach data to it. This then enables the base to have
specialized transaction impls for different transactional behaviors -
like normal and "in-frame" - and not collide with specializations in
backends (which are no longer possible).

The actual re-computation of the notifications based on the results
of tree hashing is NOT implemented yet.