Support for running pre-commit jobs. This is not used yet, but will be used to optimize the number of required identityhash re-computations during some bigger transactions that may cause an identity hash to be recomputed several times unnecessarily atm.
The transaction is no longer subclassable in backends but instead
they can attach data to it. This then enables the base to have
specialized transaction impls for different transactional behaviors -
like normal and "in-frame" - and not collide with specializations in
backends (which are no longer possible).

The actual re-computation of the notifications based on the results
of tree hashing is NOT implemented yet.

Also fixed the transactional behavior during reads that were committed
prematurely before all the data was read from the result set. This
could lead to non-optimal cache behavior in Titan and SQL backends that
would drop transactional caches too soon and would unnecessarily hit
the DB with requests that could be served locally from the cache.

We also ensure that backends are always tied to a single transaction
and always execute stuff between startTx and commit/rollback in a single
transaction.

Before this change, it could happen that a single backend could commit
stuff logically associated with 2 different transactions in a transaction
that was implicitly associated with the backend itself in tinkerpop impls.

This could lead to erratic behavior where either obsolete transactions
would be used or stuff from different transactions would appear mixed
together.

This change, although huge, doesn't change any "business" logic - it merely
changes the way transactions are associated with a backend and the
nitty-gritty details of transaction handling within base itself.