We now produce 3 kinds of hashes for different entities.
* All entity types but metdatapack have a content hash - this is just a
hash of its properties, generic properties and name and is not directly
used in the API for anything. It is NOT hierarchical, i.e. when a content
hash changes on a contained entity, the content hash of the containing
entity doesn't change.

* Identity hash works the same as before and is used for locating the
"identical" entities - the ones with the same structure, but potentially
different name or generic properties.

* Sync hash is a new kind of hash that is the union of content and identity
hashes. It is used as the basis for synchronization of inventory structure.
I.e. even a changes in generic properties or name of an entity is now
synchronized.