[HWKINVENT-192] Use mapped names when filtering by properties.
This enables the users to use the "logical" names for properties (i.e.
the ones that read the same as the JSON properties), while also keeps the
possibility to use the internal names (like __metric_data_type) that
was possible more or less accidentally before and has been relied upon by
some clients since.