Ensure a strict ordering of elements in inventory structure to make sure the hashes are unaffected by order of elements in input.
Slightly change the lazy loaded "online" inventory structure to
load and cache full nodes by default for better caching.

Also add the ability to create offline copies without copying attachments
of the structure, so that the callers can control whether or not
attachments (which have caller-defined semantics) are retained. The default
behavior of copy() is to leave out the attachments.