Optimize the pre-synchronization setup so that minimum number of backend calls is made.
When syncing everything, there's no need to merge the incoming structure
with the pre-existing one (which is read from the db).

Also when syncing parts of the structure and not using deep search to look
for entities to retain, re-use the previously loaded entities instead of
constantly reloading them from the database.